+++ src/app/@core/utils/index.ts
@@ -1,2 +1,3 @@
 export { AnalyticsService } from './analytics.service';
 export { MailService } from './mail.service';
+export { DrawService, Participant } from './draw.service';
